+++ packages/appkit/src/connectors/index.ts
@@ -3,5 +3,6 @@ export * from "./genie";
 export * from "./jobs";
 export * from "./lakebase";
 export * from "./lakebase-v1";
+export * from "./mcp";
 export * from "./sql-warehouse";
 export * from "./vector-search";
